refactor(audit): split mod.rs into submodules
Split src/audit/mod.rs (664 lines) into submodules:

- src/audit/types.rs: AuditEventType, AuditResult, AuditEvent, AuditEventBuilder
- src/audit/logger.rs: AuditLogger struct, impl, Default
- src/audit/query.rs: AuditQueryFilter, MAX_QUERY_DAYS, query_audit_events, export_audit_events
- src/audit/tests.rs: #[cfg(test)] mod tests (all unit tests)
- src/audit/mod.rs: module declarations + pub use re-exports

All 11 tests pass. Public API unchanged.

Source: issue #218

Type: refactor
